backend/backend/asgi.py
import os
os.environ.setdefault('DJANGO_SETTINGS_MODULE', 'backend.settings')

# Must call get_asgi_application() first to initialise Django apps
from django.core.asgi import get_asgi_application
django_asgi_app = get_asgi_application()

from channels.routing import ProtocolTypeRouter, URLRouter
import chat.routing
import notifications.routing
from backend.middleware import JWTAuthMiddleware

application = ProtocolTypeRouter({
    'http': django_asgi_app,
    'websocket': JWTAuthMiddleware(
        URLRouter(
            chat.routing.websocket_urlpatterns +
            notifications.routing.websocket_urlpatterns
        )
    ),
})
