backend/requirements.txt
django>=5.0
djangorestframework>=3.15
django-cors-headers>=4.3
djangorestframework-simplejwt>=5.3
Pillow>=10.0
django-channels>=4.0
channels-redis>=4.2
celery>=5.3
django-filter>=23.5
drf-spectacular>=0.27
python-decouple>=3.8
twilio>=8.0
googlemaps>=4.10
